+++ ICD.Common.Utils.Tests/Extensions/DateTimeExtensionsTest.cs
@@ -3,16 +3,16 @@
 namespace ICD.Common.Utils.Tests.Extensions
 {
 	[TestFixture]
-    public sealed class DateTimeExtensionsTest
-    {
-	    [Test]
-	    public static void ToShortTimeStringTest()
-	    {
-		    Assert.Inconclusive();
-	    }
+	public sealed class DateTimeExtensionsTest
+	{
+		[Test]
+		public void ToShortTimeStringTest()
+		{
+			Assert.Inconclusive();
+		}
 
 		[Test]
-		public static void ToLongTimeStringWithMillisecondsTest()
+		public void ToLongTimeStringWithMillisecondsTest()
 		{
 			Assert.Inconclusive();
 		}